requirements.txt
fastapi
uvicorn
wikipedia
google-generativeai